requirements.txt
openai>=1.0.0
langchain>=0.1.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
numpy>=1.24.0
PyPDF2>=3.0.0
tiktoken>=0.5.0 